S03P13A306-31 | Done | Blog update & delete Function Update

See merge request s03-webmobile1-sub3/s03p13a306!16

+++ backend/src/main/java/com/web/blog/repository/BlogRepository.java
@@ -1,16 +1,14 @@
 package com.web.blog.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.web.blog.domain.Blog;
-import com.web.blog.domain.Users;
-
 import java.util.List;
-import java.util.Optional;
 
 import javax.transaction.Transactional;
 
-public interface BlogRepository extends JpaRepository<Blog, Long> {
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.web.blog.domain.Blog;
+
+public interface BlogRepository extends JpaRepository<Blog, Integer> {
 
 	long countByManager(String manager);
